make stats.ts
